+++ gulp/styles.js
@@ -33,4 +33,17 @@ module.exports = function (options) {
         $.del([options.paths.build + '/' + options.paths.styles], done);
     });
 
+    gulp.task('extractLessVariables', function () {
+
+        return gulp.src(options.paths.src + '/**/variables.less')
+            .pipe($.debug())
+            .pipe($.less2js())
+            .pipe($.rename(function (path) {
+                path.dirname = '/config';
+                path.extname = '.json';
+            }))
+            .pipe(gulp.dest(options.paths.build))
+
+    });
+    
 };